fix: change value in input
