+++ packages/ui/src/components/DateInput/__stories__/Controlled.stories.tsx
@@ -2,6 +2,8 @@ import type { StoryFn } from '@storybook/react'
 import type { ComponentProps } from 'react'
 import { useState } from 'react'
 import { DateInput } from '..'
+import { Button } from '../../Button'
+import { Stack } from '../../Stack'
 
 export const Controlled: StoryFn<ComponentProps<typeof DateInput>> = args => {
   const [value, setValue] = useState<
@@ -9,12 +11,18 @@ export const Controlled: StoryFn<ComponentProps<typeof DateInput>> = args => {
   >(new Date('December 17, 1995 03:24:00'))
 
   return (
-    <DateInput
-      {...args}
-      label="Date"
-      value={value as Date}
-      onChange={setValue}
-    />
+    <Stack gap={2}>
+      <DateInput
+        {...args}
+        label="Date"
+        value={value as Date}
+        onChange={setValue}
+      />
+      Selected date : {value?.toString()}
+      <Button onClick={() => setValue(new Date('December 26, 1995 03:24:00'))}>
+        Set date to 1995-12-26
+      </Button>
+    </Stack>
   )
 }
