+++ packages/ui/src/components/DateInput/__stories__/Disabled.stories.tsx
@@ -6,3 +6,11 @@ Disabled.args = {
   label: 'Disabled',
   disabled: true,
 }
+
+Disabled.decorators = [
+  StoryComponent => (
+    <div style={{ height: '350px' }}>
+      <StoryComponent />
+    </div>
+  ),
+]
