+++ packages/ui/src/components/DateInput/__stories__/I18n.stories.tsx
@@ -62,6 +62,7 @@ export const I18n: StoryFn<ComponentProps<typeof DateInput>> = args => {
           setValue(v as Date)
         }}
         label={currentLocale}
+        selectsRange={false}
       />
     </div>
   )
