+++ packages/ui/src/components/DateInput/__stories__/Months.stories.tsx
@@ -1,34 +1,17 @@
-import type { StoryFn } from '@storybook/react'
-import { useState } from 'react'
-import { DateInput } from '..'
+import { Template } from './Template'
 
-export const Months: StoryFn = args => {
-  const [startDate, setStartDate] = useState<Date | null>(null)
-  const [endDate, setEndDate] = useState<Date | null>(null)
-  const onChange = (dates: [Date | null, Date | null]) => {
-    const [start, end] = dates
-    setStartDate(start)
-    setEndDate(end)
-  }
+export const Months = Template.bind({})
 
-  return (
-    <DateInput
-      label="Date"
-      startDate={startDate}
-      endDate={endDate}
-      onChange={onChange}
-      showMonthYearPicker
-      selectsRange
-      {...args}
-    />
-  )
+Months.args = {
+  label: 'Months',
+  showMonthYearPicker: true,
+  disabled: false,
 }
 
-Months.parameters = {
-  docs: {
-    description: {
-      story:
-        'Using prop `showMonthYearPicker` you can select only months and years.',
-    },
-  },
-}
+Months.decorators = [
+  StoryComponent => (
+    <div style={{ height: '350px' }}>
+      <StoryComponent />
+    </div>
+  ),
+]
