+++ packages/ui/src/components/DateInput/__stories__/Playground.stories.tsx
@@ -2,4 +2,14 @@ import { Template } from './Template'
 
 export const Playground = Template.bind({})
 
-Playground.args = Template.args
+Playground.args = {
+  placeholder: 'YYYY-MM-DD',
+}
+
+Playground.decorators = [
+  StoryComponent => (
+    <div style={{ height: '350px' }}>
+      <StoryComponent />
+    </div>
+  ),
+]
