+++ packages/ui/src/components/DateInput/__stories__/Range.stories.tsx
@@ -1,25 +1,69 @@
 import type { StoryFn } from '@storybook/react'
+import type { ComponentProps } from 'react'
 import { useState } from 'react'
 import { DateInput } from '..'
+import { Stack } from '../../Stack'
 
-export const Range: StoryFn = args => {
+const months = [
+  'January',
+  'February',
+  'March',
+  'April',
+  'May',
+  'June',
+  'July',
+  'August',
+  'September',
+  'October',
+  'November',
+  'December',
+]
+export const Range: StoryFn<ComponentProps<typeof DateInput>> = args => {
   const [startDate, setStartDate] = useState<Date | null>(null)
   const [endDate, setEndDate] = useState<Date | null>(null)
-  const onChange = (dates: [Date | null, Date | null]) => {
+
+  const [startMonth, setStartMonth] = useState<Date | null>(
+    new Date('March 2024'),
+  )
+  const [endMonth, setEndMonth] = useState<Date | null>(null)
+
+  const onChange = (dates: [Date | null, Date | null] | Date[]) => {
     const [start, end] = dates
     setStartDate(start)
     setEndDate(end)
   }
 
+  const onChangeMonth = (dates: [Date | null, Date | null] | Date[]) => {
+    const [start, end] = dates
+    setStartMonth(start)
+    setEndMonth(end)
+  }
+
   return (
-    <DateInput
-      label="Date"
-      onChange={onChange}
-      startDate={startDate}
-      endDate={endDate}
-      selectsRange
-      {...args}
-    />
+    <Stack>
+      <DateInput
+        label="Date"
+        onChange={onChange}
+        startDate={startDate}
+        endDate={endDate}
+        selectsRange
+        {...args}
+      />
+      Selected dates : {startDate?.toDateString()} - {endDate?.toDateString()}
+      <DateInput
+        label="Month"
+        onChange={onChangeMonth}
+        startDate={startMonth}
+        endDate={endMonth}
+        selectsRange
+        showMonthYearPicker
+        {...args}
+      />
+      Selected months : {startMonth ? months[startMonth.getMonth()] : null}
+      {startMonth?.getFullYear()} -
+      {endMonth ? months[endMonth.getMonth()] : null}
+      {endMonth?.getFullYear()}
+    </Stack>
   )
 }
