+++ packages/ui/src/components/DateInput/__stories__/Template.tsx
@@ -1,12 +1,12 @@
 import type { StoryFn } from '@storybook/react'
-import type { ComponentProps } from 'react'
 import { DateInput } from '..'
 
-export const Template: StoryFn<ComponentProps<typeof DateInput>> = props => (
+export const Template: StoryFn<typeof DateInput> = props => (
   <DateInput {...props} />
 )
 
 Template.args = {
   label: 'Date Input',
   required: true,
+  placeholder: 'YYYY-MM-DD',
 }
