+++ packages/ui/src/components/DateInput/__stories__/index.stories.tsx
@@ -3,26 +3,17 @@ import { DateInput } from '..'
 
 export default {
   component: DateInput,
-  decorators: [
-    Stories => (
-      <div style={{ height: 360 }}>
-        <Stories />
-      </div>
-    ),
-  ],
   title: 'Components/Data Entry/DateInput',
 } as Meta
-
 export { Playground } from './Playground.stories'
 export { Size } from './Size.stories'
-export { Exclude } from './Exclude.stories'
-export { Range } from './Range.stories'
-export { Uncontrolled } from './Uncontrolled.stories'
-export { Controlled } from './Controlled.stories'
 export { Error } from './Error.stories'
 export { Disabled } from './Disabled.stories'
 export { Required } from './Required.stories'
 export { Months } from './Months.stories'
-export { Localized } from './Localized.stories'
-export { I18n } from './I18n.stories'
+export { Exclude } from './Exclude.stories'
 export { MinMax } from './MinMax.stories'
+export { Localized } from './Localized.stories'
+export { Controlled } from './Controlled.stories'
+export { Range } from './Range.stories'
+export { I18n } from './I18n.stories'
