+++ packages/ui/src/components/DateInput/constants.ts
@@ -1,0 +1,48 @@
+export const POPUP_WIDTH = '264px'
+
+export const CURRENT_YEAR = new Date().getFullYear()
+export const CURRENT_MONTH = new Date().getMonth() + 1
+
+// Weeks displayed on calendar
+export const CALENDAR_WEEKS = 6
+
+export const DAYS_DEFAULT = {
+  Sunday: 'Su',
+  Monday: 'Mo',
+  Tuesday: 'Tu',
+  Wednesday: 'We',
+  Thursday: 'Th',
+  Friday: 'Fr',
+  Saturday: 'Sa',
+}
+
+// Calendar months names and short names
+export const MONTHS_DEFAULT = {
+  January: 'Jan',
+  February: 'Feb',
+  March: 'Mar',
+  April: 'Apr',
+  May: 'May',
+  June: 'Jun',
+  July: 'Jul',
+  August: 'Aug',
+  September: 'Sep',
+  October: 'Oct',
+  November: 'Nov',
+  December: 'Dec',
+}
+
+export const MONTHS_ARR_DEFAULT = [
+  'January',
+  'February',
+  'March',
+  'April',
+  'May',
+  'June',
+  'July',
+  'August',
+  'September',
+  'October',
+  'November',
+  'December',
+]
