--- packages/ui/src/components/DateInput/datepicker.css
@@ -1,1 +1,0 @@
-@import 'react-datepicker/dist/react-datepicker.min.css';
