+++ packages/ui/src/components/DateInput/index.tsx
@@ -1,198 +1,21 @@
-import { Global } from '@emotion/react'
 import styled from '@emotion/styled'
 import { CalendarRangeIcon } from '@ultraviolet/icons'
-import type { FocusEvent, ReactNode } from 'react'
-import { useId } from 'react'
-import type { ReactDatePickerProps } from 'react-datepicker'
-import DatePicker, { registerLocale } from 'react-datepicker'
-import { Button } from '../Button'
-import { Stack } from '../Stack'
-import { Text } from '../Text'
+import type { Locale } from 'date-fns'
+import type { ChangeEvent, FocusEvent } from 'react'
+import { useEffect, useMemo, useRef, useState } from 'react'
 import { TextInputV2 } from '../TextInputV2'
-import style from './datepicker.css?inline'
+import { type ContextProps, DateInputContext } from './Context'
+import { CalendarPopup } from './components/Popup'
+import { CURRENT_MONTH, CURRENT_YEAR } from './constants'
+import { formatValue } from './helpers'
+import { getDays, getLocalizedMonths, getMonths } from './helpersLocale'
 
-const PREFIX = '.react-datepicker'
+const Container = styled.div`
+width: 100%;`
 
-const StyledWrapper = styled.div`
-  width: 100%;
-
-  div${PREFIX}-wrapper {
-    display: block;
-  }
-  div${PREFIX}__input-container {
-    display: block;
-  }
-  div${PREFIX}__triangle {
-    display: none;
-  }
-
-  div${PREFIX}__month-container {
-    padding: ${({ theme }) => theme.space['2']};
-    width: 20rem;
-  }
-
-  ${PREFIX}-popper {
-    z-index: 1000;
-  }
-  .calendar {
-    font-family: ${({ theme }) => theme.typography.body.fontFamily};
-    border-color: ${({ theme }) => theme.colors.neutral.borderWeak};
-    background-color: ${({ theme }) =>
-      theme.colors.other.elevation.background.raised};
-    box-shadow: ${({ theme }) => theme.shadows.raised};
-
-
-    ${PREFIX}__header {
-      color: ${({ theme }) => theme.colors.neutral.text};
-      background-color: ${({ theme }) =>
-        theme.colors.other.elevation.background.raised};
-      border-bottom: none;
-      text-align: inherit;
-      display: block;
-      padding-top: 0;
-      position: inherit;
-    }
-
-    ${PREFIX}__triangle {
-      border-bottom-color: ${({ theme }) => theme.colors.neutral.backgroundWeak};
-    }
-    ${PREFIX}__month {
-      margin: 0;
-    }
-
-    ${PREFIX}__day-names {
-      margin-top: ${({ theme }) => theme.space['1']};
-      display: flex;
-      justify-content: center;
-    }
-
-    ${PREFIX}__day-name {
-      font-family: ${({ theme }) => theme.typography.bodySmallStrong.fontFamily};
-      color: ${({ theme }) => theme.colors.neutral.text};
-      font-weight: ${({ theme }) => theme.typography.bodySmallStrong.weight};
-      font-size: ${({ theme }) => theme.typography.bodySmallStrong.fontSize};
-      line-height: ${({ theme }) => theme.typography.bodySmallStrong.lineHeight};
-      text-align: center;
-      margin: ${({ theme }) => theme.space['0.5']};
-      text-transform: capitalize;
-    }
-
-    ${PREFIX}__day, ${PREFIX}__month {
-      color: ${({ theme }) => theme.colors.neutral.textWeak};
-      font-weight: ${({ theme }) => theme.typography.bodyStrong.weight};
-      font-size: ${({ theme }) => theme.typography.bodyStrong.fontSize};
-      margin-left: ${({ theme }) => theme.space['0.5']};
-      margin-right: ${({ theme }) => theme.space['0.5']};
-    }
-
-    ${PREFIX}__day {
-      width: ${({ theme }) => theme.sizing['400']};
-      height: ${({ theme }) => theme.sizing['400']};
-    }
-
-    ${PREFIX}__month-text {
-      height: ${({ theme }) => theme.sizing['400']};
-      display: inline-flex;
-      justify-content: center;
-      align-items: center;
-    }
-
-    ${PREFIX}__day--outside-month {
-      color: ${({ theme }) => theme.colors.neutral.textDisabled};
-      font-weight: ${({ theme }) => theme.typography.bodyStrong.weight};
-      font-size: ${({ theme }) => theme.typography.bodyStrong.fontSize};
-    }
-
-    ${PREFIX}__day--selected, ${PREFIX}__month-text--selected {
-      color: ${({ theme }) => theme.colors.primary.textStrong};
-      background-color: ${({ theme }) => theme.colors.primary.backgroundStrong};
-
-      &[aria-disabled="true"],
-      &:disabled {
-        color: ${({ theme }) => theme.colors.primary.textStrongDisabled};
-        background-color: ${({ theme }) =>
-          theme.colors.primary.backgroundStrongDisabled};
-      }
-    }
-
-    ${PREFIX}__day--in-selecting-range,  ${PREFIX}__month-text--in-selecting-range {
-      color: ${({ theme }) => theme.colors.primary.text};
-      background-color: ${({ theme }) => theme.colors.primary.background};
-
-      &[aria-disabled="true"],
-      &:disabled {
-        color: ${({ theme }) => theme.colors.primary.textDisabled};
-        background-color: ${({ theme }) =>
-          theme.colors.primary.backgroundDisabled};
-      }
-    }
-
-    ${PREFIX}__day--in-range, ${PREFIX}__month-text--in-range {
-      color: ${({ theme }) => theme.colors.primary.text};
-      background-color: ${({ theme }) => theme.colors.primary.background};
-
-      &[aria-disabled="true"],
-      &:disabled {
-        color: ${({ theme }) => theme.colors.primary.textDisabled};
-        background-color: ${({ theme }) =>
-          theme.colors.primary.backgroundDisabled};
-      }
-    }
-
-    ${PREFIX}__day--range-start, ${PREFIX}__month-text--range-start {
-      color: ${({ theme }) => theme.colors.primary.textStrong};
-      background-color: ${({ theme }) => theme.colors.primary.backgroundStrong};
-
-      &[aria-disabled="true"],
-      &:disabled {
-        color: ${({ theme }) => theme.colors.primary.textStrongDisabled};
-        background-color: ${({ theme }) =>
-          theme.colors.primary.backgroundStrongDisabled};
-      }
-    }
-
-    ${PREFIX}__day--range-end, ${PREFIX}__month-text--range-end {
-      color: ${({ theme }) => theme.colors.primary.textStrong};
-      background-color: ${({ theme }) => theme.colors.primary.backgroundStrong};
-
-      &[aria-disabled="true"],
-      &:disabled {
-        color: ${({ theme }) => theme.colors.primary.textStrongDisabled};
-        background-color: ${({ theme }) =>
-          theme.colors.primary.backgroundStrongDisabled};
-      }
-    }
-
-    ${PREFIX}__day--keyboard-selected, ${PREFIX}__month-text--keyboard-selected {
-      color: ${({ theme }) => theme.colors.primary.textStrong};
-      background-color: ${({ theme }) => theme.colors.primary.backgroundStrong};
-    }
-
-    ${PREFIX}__day:hover, ${PREFIX}__month-text:hover {
-      color: ${({ theme }) => theme.colors.neutral.textHover};
-      background-color: ${({ theme }) => theme.colors.neutral.backgroundHover};
-    }
-
-    ${PREFIX}__day--disabled, ${PREFIX}__month-text--disabled {
-      color: ${({ theme }) => theme.colors.neutral.textDisabled};
-    }
-
-    ${PREFIX}__day--disabled:hover, ${PREFIX}__month-text--disabled:hover {
-      color: ${({ theme }) => theme.colors.neutral.textDisabled};
-      background-color: transparent;
-    }
-  }
-`
-
-const StyledText = styled(Text)`
-  text-transform: capitalize;
-`
-
-type DateInputProps = Pick<
-  ReactDatePickerProps<boolean | undefined, boolean>,
-  'locale' | 'onChange'
-> & {
+type DateInputProps<IsRange extends undefined | boolean = false> = {
   autoFocus?: boolean
+  locale?: string | Locale
   disabled?: boolean
   maxDate?: Date | null
   minDate?: Date | null
@@ -201,43 +24,46 @@ type DateInputProps = Pick<
   onFocus?: (event: FocusEvent<HTMLInputElement>) => void
   error?: string
   required?: boolean
-  format?: (value?: Date | string) => string | undefined
+  format?: (value?: Date) => string | undefined
   /**
    * Label of the field
    */
   label?: string
-  value?: Date | string | [Date | null, Date | null]
+  value?: Date | string | null
   className?: string
   'data-testid'?: string
-  selectsRange?: boolean
-  startDate?: Date | null
-  endDate?: Date | null
   excludeDates?: Date[]
   id?: string
-  labelDescription?: ReactNode
+  labelDescription?: string
   success?: string | boolean
   helper?: string
   size?: 'small' | 'medium' | 'large'
   readOnly?: boolean
   tooltip?: string
   showMonthYearPicker?: boolean
+  placeholder?: string
+  startDate?: Date | null
+  endDate?: Date | null
+  selectsRange?: IsRange
+  onChange?: IsRange extends true
+    ? (
+        date: Date[] | [Date | null, Date | null],
+        event: React.SyntheticEvent | undefined,
+      ) => void
+    : (date: Date | null, event: React.SyntheticEvent | undefined) => void
 }
 
-const DEFAULT_FORMAT: DateInputProps['format'] = value =>
-  value instanceof Date ? value.toISOString() : value
-
 /**
- * DateInput is a wrapper around react-datepicker that provides a consistent look and feel with the rest of the Ultraviolet UI.
- * See https://reactdatepicker.com/ for more information.
+ * DateInput can be used to select a specific date
  */
-export const DateInput = ({
+export const DateInput = <IsRange extends undefined | boolean>({
   autoFocus = false,
   disabled = false,
   error,
-  format = DEFAULT_FORMAT,
+  format,
   label,
   labelDescription,
-  locale,
+  locale = 'en-US',
   maxDate,
   minDate,
   startDate,
@@ -245,11 +71,11 @@ export const DateInput = ({
   name,
   onBlur,
   onChange,
+  placeholder,
   onFocus,
   required = false,
   excludeDates,
   value,
-  selectsRange,
   className,
   id,
   success,
@@ -257,125 +83,207 @@ export const DateInput = ({
   size = 'large',
   readOnly = false,
   tooltip,
-  showMonthYearPicker,
+  selectsRange = false,
+  showMonthYearPicker = false,
   'data-testid': dataTestId,
-}: DateInputProps) => {
-  const uniqueId = useId()
-  const localId = id ?? uniqueId
+}: DateInputProps<IsRange>) => {
+  const defaultMonthToShow = useMemo(() => {
+    if (value) return new Date(value).getMonth() + 1
+    if (startDate && selectsRange) return startDate.getMonth() + 1
+    if (endDate && selectsRange) return endDate.getMonth() + 1
 
-  // Linked to: https://github.com/Hacker0x01/react-datepicker/issues/3834
-  const ReactDatePicker =
-    (DatePicker as unknown as { default: typeof DatePicker }).default ??
-    DatePicker
+    return CURRENT_MONTH
+  }, [endDate, selectsRange, startDate, value])
 
-  const localeCode =
-    (typeof locale === 'string' ? locale : locale?.code) ?? 'en-GB'
+  const defaultYearToShow = useMemo(() => {
+    if (value) return new Date(value).getFullYear()
+    if (startDate && selectsRange) return startDate.getFullYear()
+    if (endDate && selectsRange) return endDate.getFullYear()
 
-  if (typeof locale === 'object') {
-    registerLocale(localeCode, locale)
+    return CURRENT_YEAR
+  }, [endDate, selectsRange, startDate, value])
+
+  const [computedValue, setValue] = useState(
+    value && !selectsRange ? new Date(value) : null,
+  )
+  const [computedRange, setRange] = useState({
+    start: startDate ?? null,
+    end: endDate ?? null,
+  })
+  const [isPopupVisible, setVisible] = useState(false)
+  const [monthToShow, setMonthToShow] = useState(defaultMonthToShow)
+  const [yearToShow, setYearToShow] = useState(defaultYearToShow)
+  const [inputValue, setInputValue] = useState(
+    formatValue(
+      computedValue,
+      computedRange,
+      showMonthYearPicker,
+      selectsRange,
+      format,
+    ),
+  )
+  const refInput = useRef<HTMLInputElement>(null)
+  const MONTHS = getMonths(locale)
+  const DAYS = getDays(locale)
+  const MONTHS_ARR = getLocalizedMonths(locale)
+
+  const valueContext = useMemo(
+    () =>
+      ({
+        showMonthYearPicker,
+        disabled,
+        value: computedValue,
+        range: computedRange,
+        setRange,
+        setValue,
+        monthToShow,
+        yearToShow,
+        setMonthToShow,
+        setYearToShow,
+        excludeDates,
+        maxDate,
+        minDate,
+        MONTHS,
+        MONTHS_ARR,
+        DAYS,
+        onChange,
+        selectsRange,
+        format,
+        setInputValue,
+      }) as ContextProps,
+    [
+      showMonthYearPicker,
+      disabled,
+      selectsRange,
+      computedValue,
+      computedRange,
+      monthToShow,
+      yearToShow,
+      excludeDates,
+      maxDate,
+      minDate,
+      MONTHS,
+      MONTHS_ARR,
+      DAYS,
+      onChange,
+      format,
+      setInputValue,
+    ],
+  )
+
+  useEffect(() => {
+    if (value && !selectsRange) {
+      setValue(new Date(value))
+      setInputValue(
+        formatValue(
+          new Date(value),
+          null,
+          showMonthYearPicker,
+          selectsRange,
+          format,
+        ),
+      )
+    }
+    if (selectsRange) {
+      setRange({
+        start: startDate ?? computedRange.start,
+        end: endDate ?? computedRange.end,
+      })
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [endDate, startDate, value])
+
+  const manageOnChange = (event: ChangeEvent<HTMLInputElement>) => {
+    const newValue = event.currentTarget.value
+
+    if (selectsRange) {
+      const [startDateInput, endDateInput] = newValue
+        .split('–')
+        .map(val => new Date(val))
+
+      const computedNewRange: [Date | null, Date | null] = [
+        Number.isNaN(startDateInput.getTime())
+          ? computedRange.start
+          : startDateInput,
+        Number.isNaN(endDateInput.getTime()) ? computedRange.end : endDateInput,
+      ]
+
+      setRange({ start: computedNewRange[0], end: computedNewRange[1] })
+      setInputValue(newValue)
+
+      if (computedNewRange[0]) {
+        setMonthToShow(computedNewRange[0].getMonth() + 1)
+        setYearToShow(computedNewRange[0].getFullYear())
+      }
+      // TypeScript fails to automatically get the correct type of onChange here
+      ;(
+        onChange as (
+          date: Date[] | [Date | null, Date | null],
+          event: React.SyntheticEvent | undefined,
+        ) => void
+      )?.(computedNewRange, event)
+    } else {
+      const computedDate = new Date(newValue)
+      setInputValue(newValue)
+
+      if (Date.parse(newValue)) {
+        setValue(computedDate)
+        setMonthToShow(computedDate.getMonth() + 1)
+        setYearToShow(computedDate.getFullYear())
+
+        // TypeScript fails to automatically get the correct type of onChange here
+        ;(
+          onChange as (date: Date | null, event?: React.SyntheticEvent) => void
+        )?.(computedDate, event)
+      }
+    }
   }
 
-  const valueStart = `${
-    startDate !== undefined && startDate !== null
-      ? `${format(startDate)} -`
-      : ''
-  }`
-  const valueEnd = `${
-    endDate !== undefined && endDate !== null ? format(endDate) : ''
-  }`
-
-  const valueFormat = selectsRange
-    ? `${valueStart} ${valueEnd}`
-    : format(value as Date)
-
   return (
-    <>
-      <Global styles={style} />
-      <StyledWrapper>
-        <ReactDatePicker
-          required={required}
-          data-testid={dataTestId}
-          className={className}
-          autoFocus={autoFocus}
-          fixedHeight
-          name={name}
-          locale={localeCode}
-          onBlur={onBlur}
-          onChange={onChange}
-          onFocus={onFocus}
-          selected={
-            value && !selectsRange ? new Date(value as Date) : undefined
-          }
-          selectsRange={selectsRange}
-          excludeDates={excludeDates}
-          showPopperArrow={false}
-          popperPlacement="bottom-start"
-          customInput={
-            <TextInputV2
-              error={error}
-              success={success}
-              helper={helper}
-              id={localId}
-              label={label}
-              labelDescription={labelDescription}
-              value={valueFormat}
-              disabled={disabled}
-              size={size}
-              suffix={
-                <CalendarRangeIcon sentiment="neutral" disabled={disabled} />
-              }
-              readOnly={readOnly}
-              tooltip={tooltip}
-            />
-          }
-          disabled={disabled}
-          calendarClassName="calendar"
-          minDate={minDate}
-          maxDate={maxDate}
-          startDate={startDate}
-          endDate={endDate}
-          showMonthYearPicker={showMonthYearPicker}
-          dateFormat={showMonthYearPicker ? 'MM/yyyy' : undefined}
-          renderCustomHeader={({
-            date,
-            /* eslint-disable-next-line @typescript-eslint/unbound-method */
-            decreaseMonth,
-            /* eslint-disable-next-line @typescript-eslint/unbound-method */
-            increaseMonth,
-            prevMonthButtonDisabled,
-            nextMonthButtonDisabled,
-          }) => (
-            <Stack
-              direction="row"
-              alignItems="center"
-              justifyContent="space-between"
-            >
-              <Button
-                size="small"
-                icon="arrow-left"
+    <DateInputContext.Provider value={valueContext}>
+      <Container
+        id={id}
+        className={className}
+        data-testid={dataTestId}
+        onBlur={onBlur}
+        onFocus={onFocus}
+        onClick={() => {
+          if (!isPopupVisible) setVisible(true)
+        }}
+      >
+        <CalendarPopup
+          visible={isPopupVisible}
+          setVisible={setVisible}
+          refInput={refInput}
+        >
+          <TextInputV2
+            label={label}
+            placeholder={placeholder}
+            value={inputValue}
+            required={required}
+            error={error}
+            success={success}
+            readOnly={readOnly}
+            disabled={disabled}
+            size={size}
+            autoFocus={autoFocus}
+            helper={helper}
+            labelDescription={labelDescription}
+            name={name}
+            suffix={
+              <CalendarRangeIcon
+                size="large"
                 sentiment="neutral"
-                variant="ghost"
-                onClick={decreaseMonth}
-                disabled={prevMonthButtonDisabled}
+                disabled={disabled}
               />
-              <StyledText variant="bodyStrong" as="p">
-                {new Date(date).toLocaleString(localeCode, {
-                  month: 'long',
-                  year: 'numeric',
-                })}
-              </StyledText>
-              <Button
-                size="small"
-                icon="arrow-right"
-                sentiment="neutral"
-                variant="ghost"
-                onClick={increaseMonth}
-                disabled={nextMonthButtonDisabled}
-              />
-            </Stack>
-          )}
-        />
-      </StyledWrapper>
-    </>
+            }
+            ref={refInput}
+            tooltip={tooltip}
+            autoComplete="false"
+            onChange={manageOnChange}
+          />
+        </CalendarPopup>
+      </Container>
+    </DateInputContext.Provider>
   )
 }
