--- packages/ui/src/components/DateInputV2/__stories__/Disabled.stories.tsx
@@ -1,16 +1,0 @@
-import { Template } from './Template'
-
-export const Disabled = Template.bind({})
-
-Disabled.args = {
-  label: 'Disabled',
-  disabled: true,
-}
-
-Disabled.decorators = [
-  StoryComponent => (
-    <div style={{ height: '350px' }}>
-      <StoryComponent />
-    </div>
-  ),
-]
