--- packages/ui/src/components/DateInputV2/__stories__/Months.stories.tsx
@@ -1,17 +1,0 @@
-import { Template } from './Template'
-
-export const Months = Template.bind({})
-
-Months.args = {
-  label: 'Months',
-  showMonthYearPicker: true,
-  disabled: false,
-}
-
-Months.decorators = [
-  StoryComponent => (
-    <div style={{ height: '350px' }}>
-      <StoryComponent />
-    </div>
-  ),
-]
