--- packages/ui/src/components/DateInputV2/__stories__/Playground.stories.tsx
@@ -1,15 +1,0 @@
-import { Template } from './Template'
-
-export const Playground = Template.bind({})
-
-Playground.args = {
-  placeholder: 'YYYY-MM-DD',
-}
-
-Playground.decorators = [
-  StoryComponent => (
-    <div style={{ height: '350px' }}>
-      <StoryComponent />
-    </div>
-  ),
-]
