--- packages/ui/src/components/DateInputV2/__stories__/Required.stories.tsx
@@ -1,8 +1,0 @@
-import { Template } from './Template'
-
-export const Required = Template.bind({})
-
-Required.args = {
-  label: 'Required',
-  required: true,
-}
