--- packages/ui/src/components/DateInputV2/__stories__/index.stories.tsx
@@ -1,19 +1,0 @@
-import type { Meta } from '@storybook/react'
-import { DateInputV2 } from '..'
-
-export default {
-  component: DateInputV2,
-  title: 'Components/Data Entry/DateInputV2',
-} as Meta
-export { Playground } from './Playground.stories'
-export { Size } from './Size.stories'
-export { Error } from './Error.stories'
-export { Disabled } from './Disabled.stories'
-export { Required } from './Required.stories'
-export { Months } from './Months.stories'
-export { Exclude } from './Exclude.stories'
-export { MinMax } from './MinMax.stories'
-export { Localized } from './Localized.stories'
-export { Controlled } from './Controlled.stories'
-export { Range } from './Range.stories'
-export { I18n } from './I18n.stories'
