+++ packages/ui/src/components/index.ts
@@ -16,7 +16,6 @@ export { CheckboxGroup, CheckboxGroupCheckbox } from './CheckboxGroup'
 export { Chip } from './Chip'
 export { CopyButton } from './CopyButton'
 export { DateInput } from './DateInput'
-export { DateInputV2 } from './DateInputV2'
 export { Dialog } from './Dialog'
 export { EmptyState } from './EmptyState'
 export { Expandable } from './Expandable'
